+++ lib/fields/date_field.dart
@@ -4,7 +4,8 @@ import '../abstract_component.dart';
 import '../mixins.dart';
 import '../utils.dart';
 
-class DateField extends Component with Field<DateTime>, MixinDisable {
+class DateField extends Component
+    with Field<DateTime>, UrlStateComponent<DateTime>, MixinDisable {
   DateField() : super('DateField') {
     nodeRoot.onChange.listen((event) {
       try {
@@ -42,6 +43,14 @@ class DateField extends Component with Field<DateTime>, MixinDisable {
     final newValue = formatDate(value);
     nodeRoot.value = newValue;
   }
+
+  @override
+  String get urlState => getStringValue();
+
+  @override
+  set urlState(String newValue) => value = DateTime.fromMillisecondsSinceEpoch(
+      (nodeRoot.valueAsNumber ?? 0).ceil(),
+      isUtc: true);
 
   String getStringValue() => formatDate(value);