+++ lib/simple_utils.dart
@@ -1,9 +1,0 @@
-
-List<String> listFromJson(List listObject) {
-  final ret = <String>[];
-  for (final val in listObject) {
-    ret.add(val.toString());
-  }
-
-  return ret;
-}